Adjusting the way integration tests work to allow for more
