+++ src/__tests__/integratation-tests.integration.test.ts
@@ -1,0 +1,8 @@
+import { describe } from "vitest";
+import { tableSuite } from "./table-test";
+import { baseRepositorySuite } from "../repository/__tests__/base-repository-test";
+
+describe("Running Integration Tests", () => {
+  tableSuite();
+  baseRepositorySuite();
+});
