+++ vitest.config.ts
@@ -1,8 +1,8 @@
 import { defineConfig } from "vitest/config";
 
 export default defineConfig({
-	test: {
-		include: ["./src/**/*.test.ts"],
-		exclude: ["./src/__tests__/*.test.ts"],
-	},
+  test: {
+    include: ["./src/**/*.test.ts"],
+    exclude: ["./src/**/*.integration.test.ts"],
+  },
 });
